Make the ingestion and clock test helpers available to every backend
