+++ src/ServiceControl.Persistence.Tests/IPersistenceTestsContext.cs
@@ -1,6 +1,7 @@
 #nullable enable
 namespace ServiceControl.Persistence.Tests;
 
+using System;
 using System.Threading.Tasks;
 using MessageFailures;
 using Microsoft.Extensions.Hosting;
@@ -14,6 +15,11 @@ public interface IPersistenceTestsContext
     Task TearDown();
 
     Task CompleteDatabaseOperation();
+
+    /// <summary>
+    /// Move the clock the persister stamps its own timestamps from
+    /// </summary>
+    void AdvanceClock(TimeSpan by);
 
     PersistenceSettings PersistenceSettings { get; }
