+++ src/ServiceControl.Persistence.Tests/PersistenceTestBase.cs
@@ -83,6 +83,8 @@ public abstract class PersistenceTestBase
         return failedMessage;
     }
 
+    protected void AdvanceClock(TimeSpan by) => PersistenceTestsContext.AdvanceClock(by);
+
     protected static async Task WaitUntil(Func<Task<bool>> conditionChecker, string condition, TimeSpan timeout = default)
     {
         timeout = timeout == default ? TimeSpan.FromSeconds(10) : timeout;
